+++ config/settings.php
@@ -75,5 +75,8 @@ return [
         'own_rooms_pagination_page_size' => env('OWN_ROOMS_PAGINATION_PAGE_SIZE',5),
         'pagination_page_size' => env('DEFAULT_PAGINATION_PAGE_SIZE', 15),
         'log_attendance' => env('LOG_ATTENDANCE', false),
+        'banner' => [
+            'enabled' => false
+        ]
     ]
 ];