+++ resources/js/views/settings/Application.vue
@@ -257,6 +257,278 @@
         </b-form-group>
 
         <hr>
+
+        <!-- Banner -->
+        <b-form-group
+          label-class="font-weight-bold"
+          :state='fieldState("banner")'
+          :label="$t('settings.application.banner.title')"
+          ref='banner-form-group'
+        >
+          <b-form-group
+            :state='fieldState("banner.enabled")'
+          >
+            <b-form-checkbox
+              id='banner-enabled'
+              v-model='settings.banner.enabled'
+              :state="fieldState('banner.enabled')"
+              :disabled='isBusy || viewOnly || !loaded'
+              switch
+            >
+              {{ $t('settings.application.banner.enabled') }}
+            </b-form-checkbox>
+
+            <template slot='invalid-feedback'>
+              <div v-html="fieldError('banner.enabled')"></div>
+            </template>
+          </b-form-group>
+
+          <banner
+            class='mt-4'
+            v-if='settings.banner.enabled'
+            :background='settings.banner.background'
+            :color='settings.banner.color'
+            :enabled='settings.banner.enabled'
+            :icon='settings.banner.icon'
+            :link='settings.banner.link'
+            :message='settings.banner.message'
+            :title='settings.banner.title'
+            :link-target='settings.banner.link_target'
+            :link-text='settings.banner.link_text'
+            :link-style='settings.banner.link_style'
+          ></banner>
+
+          <b-row
+            cols='12'
+            class='mt-4'
+            v-if='settings.banner.enabled'
+          >
+            <b-col sm='6'>
+              <b-form-group
+                label-for='banner-title-input'
+                :state='fieldState("banner.title")'
+                :label="$t('settings.application.banner.bannerTitle')"
+              >
+                <b-form-input
+                  id='banner-title-input'
+                  v-model='settings.banner.title'
+                  type='text'
+                  :disabled='isBusy || viewOnly || !loaded'
+                  :state='fieldState("banner.title")'
+                ></b-form-input>
+
+                <template slot='invalid-feedback'>
+                  <div v-html="fieldError('banner.title')"></div>
+                </template>
+              </b-form-group>
+            </b-col>
+            <b-col sm='6'>
+              <b-form-group
+                label-for='banner-icon-input'
+                :state='fieldState("banner.icon")'
+                :description="$t('settings.application.banner.iconDescription')"
+                :label="$t('settings.application.banner.icon')"
+              >
+                <b-form-input
+                  id='banner-icon-input'
+                  v-model='settings.banner.icon'
+                  type='text'
+                  :disabled='isBusy || viewOnly || !loaded'
+                  :state='fieldState("banner.icon")'
+                ></b-form-input>
+
+                <template slot='invalid-feedback'>
+                  <div v-html="fieldError('banner.icon')"></div>
+                </template>
+              </b-form-group>
+            </b-col>
+          </b-row>
+
+          <b-form-group
+            label-for='banner-message-input'
+            :state='fieldState("banner.message")'
+            :label="$t('settings.application.banner.message')"
+            v-if='settings.banner.enabled'
+          >
+            <b-form-textarea
+              id='banner-message-input'
+              v-model='settings.banner.message'
+              rows='3'
+              :disabled='isBusy || viewOnly || !loaded'
+              :state='fieldState("banner.message")'
+            ></b-form-textarea>
+
+            <template slot='invalid-feedback'>
+              <div v-html="fieldError('banner.message')"></div>
+            </template>
+          </b-form-group>
+
+          <b-row
+            cols='12'
+            v-if='settings.banner.enabled'
+          >
+            <b-col sm='6'>
+              <b-form-group
+                label-for='banner-link-input'
+                :state='fieldState("banner.link")'
+                :label="$t('settings.application.banner.link')"
+              >
+                <b-form-input
+                  id='banner-link-input'
+                  v-model='settings.banner.link'
+                  type='text'
+                  :disabled='isBusy || viewOnly || !loaded'
+                  :state='fieldState("banner.link")'
+                ></b-form-input>
+
+                <template slot='invalid-feedback'>
+                  <div v-html="fieldError('banner.link')"></div>
+                </template>
+              </b-form-group>
+            </b-col>
+            <b-col sm='6'>
+              <b-form-group
+                label-for='banner-link-text-input'
+                :state='fieldState("banner.link_text")'
+                :label="$t('settings.application.banner.link_text')"
+              >
+                <b-form-input
+                  id='banner-link-text-input'
+                  v-model='settings.banner.link_text'
+                  type='text'
+                  :disabled='isBusy || viewOnly || !loaded'
+                  :state='fieldState("banner.link_text")'
+                ></b-form-input>
+
+                <template slot='invalid-feedback'>
+                  <div v-html="fieldError('banner.link_text')"></div>
+                </template>
+              </b-form-group>
+            </b-col>
+          </b-row>
+
+          <b-row
+            cols='12'
+            v-if='settings.banner.enabled'
+          >
+            <b-col sm='6'>
+              <b-form-group
+                label-for='banner-link-style-input'
+                :state='fieldState("banner.link_style")'
+                :label="$t('settings.application.banner.link_style')"
+              >
+                <b-form-select
+                  id='banner-link-style-input'
+                  :disabled='isBusy || viewOnly || !loaded'
+                  :state='fieldState("banner.link_style")'
+                  v-model='settings.banner.link_style'
+                  :options='linkBtnStyles'
+                >
+                  <template #first>
+                    <b-form-select-option :value='null' disabled>
+                      {{ $t('settings.application.banner.selectLinkStyle') }}
+                    </b-form-select-option>
+                  </template>
+                </b-form-select>
+
+                <template slot='invalid-feedback'>
+                  <div v-html="fieldError('banner.link_style')"></div>
+                </template>
+              </b-form-group>
+            </b-col>
+            <b-col sm='6'>
+              <b-form-group
+                label-for='banner-link-target-input'
+                :state='fieldState("banner.link_target")'
+                :label="$t('settings.application.banner.link_target')"
+              >
+                <b-form-select
+                  id='banner-link-target-input'
+                  :disabled='isBusy || viewOnly || !loaded'
+                  :state='fieldState("banner.link_target")'
+                  v-model='settings.banner.link_target'
+                  :options='linkTargets'
+                >
+                  <template #first>
+                    <b-form-select-option :value='null' disabled>
+                      {{ $t('settings.application.banner.selectLinkTarget') }}
+                    </b-form-select-option>
+                  </template>
+                </b-form-select>
+
+                <template slot='invalid-feedback'>
+                  <div v-html="fieldError('banner.link_target')"></div>
+                </template>
+              </b-form-group>
+            </b-col>
+          </b-row>
+
+          <b-row
+            cols='12'
+            v-if='settings.banner.enabled'
+          >
+            <b-col sm='6'>
+              <b-form-group
+                label-for='banner-color-input'
+                :state='fieldState("banner.color")'
+                :label="$t('settings.application.banner.color')"
+              >
+                <v-swatches
+                  class='my-2'
+                  :disabled='isBusy || !loaded || viewOnly'
+                  :swatch-style="{ borderRadius: '0px', marginBottom: '11px' }"
+                  :swatches='colorSwatches'
+                  v-model='settings.banner.color'
+                  inline></v-swatches>
+                <b-form-text>{{ $t('settings.roomTypes.customColor') }}</b-form-text>
+                <b-form-input
+                  id='banner-color-input'
+                  type='text'
+                  v-model='settings.banner.color'
+                  :state='fieldState("banner.color")'
+                  :disabled='isBusy || !loaded || viewOnly'
+                ></b-form-input>
+
+                <template slot='invalid-feedback'>
+                  <div v-html="fieldError('banner.color')"></div>
+                </template>
+              </b-form-group>
+            </b-col>
+            <b-col sm='6'>
+              <b-form-group
+                label-for='banner-background-input'
+                :state='fieldState("banner.background")'
+                :label="$t('settings.application.banner.background')"
+              >
+                <v-swatches
+                  class='my-2'
+                  :disabled='isBusy || !loaded || viewOnly'
+                  :swatch-style="{ borderRadius: '0px', marginBottom: '11px' }"
+                  :swatches='backgroundSwatches'
+                  v-model='settings.banner.background'
+                  inline></v-swatches>
+                <b-form-text>{{ $t('settings.roomTypes.customColor') }}</b-form-text>
+                <b-form-input
+                  id='banner-background-input'
+                  type='text'
+                  v-model='settings.banner.background'
+                  :state='fieldState("banner.background")'
+                  :disabled='isBusy || !loaded || viewOnly'
+                ></b-form-input>
+
+                <template slot='invalid-feedback'>
+                  <div v-html="fieldError('banner.background')"></div>
+                </template>
+              </b-form-group>
+            </b-col>
+          </b-row>
+
+          <template slot='invalid-feedback'>
+            <div v-html="fieldError('banner')"></div>
+          </template>
+        </b-form-group>
+
+        <hr>
         <div class="clearfix">
           <b-button id="application-save-button"
                     class="float-right mr-1 mb-1"
@@ -277,8 +549,12 @@ import Base from '../../api/base';
 import FieldErrors from '../../mixins/FieldErrors';
 import env from '../../env';
 import PermissionService from '../../services/PermissionService';
+import Banner from '../../components/Banner';
+import VSwatches from 'vue-swatches';
+import 'vue-swatches/dist/vue-swatches.css';
 
 export default {
+  components: { Banner, VSwatches },
   mixins: [FieldErrors],
 
   data () {
@@ -290,8 +566,14 @@ export default {
       uploadFaviconFile: null,
       uploadFaviconFileSrc: null,
       isBusy: false,
-      settings: {},
-      errors: {}
+      settings: {
+        banner: {},
+        link_btn_styles: [],
+        link_targets: []
+      },
+      errors: {},
+      colorSwatches: ['#fff', '#000'],
+      backgroundSwatches: ['#4a5c66', '#80ba24', '#9C132E', '#F4AA00', '#00B8E4', '#002878']
     };
   },
   methods: {
@@ -300,7 +582,7 @@ export default {
      */
     getSettings () {
       this.isBusy = true;
-      Base.call('settings')
+      Base.call('settings/all')
         .then(response => {
           this.settings = response.data.data;
           this.roomLimitMode = (this.settings.room_limit === -1 ? 'unlimited' : 'custom');
@@ -345,6 +627,21 @@ export default {
       formData.append('room_limit', this.settings.room_limit);
       formData.append('pagination_page_size', this.settings.pagination_page_size);
       formData.append('own_rooms_pagination_page_size', this.settings.own_rooms_pagination_page_size);
+
+      Object.keys(this.settings.banner).forEach(key => {
+        let val = this.settings.banner[key];
+
+        // Since the FormData always strings boolean and empty values must be
+        // changed so that they can be handled correctly by the backend.
+        if (typeof (val) === 'boolean') {
+          val = val ? 1 : 0;
+        } else if (!val) {
+          val = '';
+        }
+
+        formData.append(`banner[${key}]`, val);
+      });
+
       formData.append('_method', 'PUT');
 
       Base.call('settings',
@@ -383,10 +680,10 @@ export default {
     roomLimitModeChanged (value) {
       switch (value) {
         case 'unlimited':
-          this.settings.room_limit = -1;
+          this.$set(this.settings, 'room_limit', -1);
           break;
         case 'custom':
-          this.settings.room_limit = 0;
+          this.$set(this.settings, 'room_limit', 0);
           break;
       }
     },
@@ -409,6 +706,17 @@ export default {
     this.getSettings();
   },
   computed: {
+    linkBtnStyles () {
+      return (this.settings.link_btn_styles || []).map((style) => {
+        return { value: style, text: this.$t(`app.buttonStyles.${style}`) };
+      });
+    },
+
+    linkTargets () {
+      return (this.settings.link_targets || []).map((target) => {
+        return { value: target, text: this.$t(`app.linkTargets.${target}`) };
+      });
+    },
 
     /**
      * Check if user is only allowed to read settings